EvtxExtract/requirements.txt
python-evtx>=0.7.4
lxml>=4.9
